app can supports only specific mime types

+++ routes/viewer.js
@@ -5,6 +5,8 @@ var fs = require("fs");
 var Q = require("q");
 var config = require("config");
 var deleteFile = require("../lib/deleter");
+var mimeTypes = require("../lib/mimeTypes");
+var logger = require("../lib/logger");
 
 module.exports = router;
 
@@ -12,15 +14,23 @@ router.use(function(request, response, next){
     var url = request.query.url;
     request.tempFile = config.get("outdir") + "\\" + request.flowUUID;
     var fileWriteStream = fs.createWriteStream(request.tempFile);
-    httpRequest(url).pipe(fileWriteStream);
-    next();
+    httpRequest(url, function(error, res, body){
+        if(mimeTypes.isSupportedType(res.headers['content-type'])){
+            next();
+        }
+        else{
+            logger.warn("mime type " + res.headers['content-type'] + " is not supported");
+            cleanup(request);
+            response.status(415).json({message:"mime type " + res.headers['content-type'] + " is not supported"});
+        }
+    }).pipe(fileWriteStream);
+    //next();
 });
 
 router.use(function(request, response, next){
     converter(request.tempFile)
         .then(function(pdfFile){
             request.pdfFile = pdfFile;
-            deleteFile(request.tempFile);
             next();
         })
 });
@@ -30,7 +40,16 @@ router.use(function(request, response, next){
     response.setHeader('Content-type', 'application/pdf');
     fs.createReadStream(request.pdfFile).pipe(response);
     response.on("finish", function(){
-        deleteFile(request.pdfFile);
+        cleanup(request);
     });
 });
+
+function cleanup(request){
+    if(request.tempFile){
+        deleteFile(request.tempFile);
+    }
+    if(request.pdfFile){
+        deleteFile(request.pdfFile);
+    }
+}
